+++ ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: kle-rest <kle-rest@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 15:54:56 by kle-rest          #+#    #+#             */
-/*   Updated: 2022/11/09 13:39:28 by kle-rest         ###   ########.fr       */
+/*   Updated: 2022/11/12 16:21:33 by kle-rest         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include "libft.h"
 
-int	ft_strlcpy(char *dest, const char *src, size_t n)
+size_t	ft_strlcpy(char *dest, const char *src, size_t n)
 {
-	size_t i;
+	size_t	i;
 
 	i = 0;
 	if (n <= 0)
@@ -28,7 +28,6 @@ int	ft_strlcpy(char *dest, const char *src, size_t n)
 	dest[i] = '\0';
 	while (src[i])
 		i++;
-//	printf("%s, %ld", dest, i);
 	return (i);
 }
 /*
